feat(settings): App configuration in individual files (issue 91)
The initial ShinyProxySpecMainProvider was able to load Add settings from application.yml.
The new version will delegate the loading task to ShinyProxySpecMainProvider and ShinyProxySpecExternalProvider
ShinyProxyspecMainProvider does the same then the initial ShinyProxySpecMainProvider, ie. it loads settings from application.yml.
ShinyProxySpecExternalProvider will try to load files from a specific path. This path can be set in applicaton.yml using the new proxy.specs-external-path attribute.
The attribute used in the individual settings files are quite the same (using camel case) :
- id, displayName, description, containerCmd, logoUrl, accessGroups, ...
